Do not pass dragover to children
It seems to be more performant.

API-wise it makes sense that withDraggable takes care of calling
children's dragstart/dragend event handlers as it needs them
to be executed after its own. The reason for this is that the
children's event handlers could modify the DOM element
withDraggable clones, messing up with its behavior.
For example, if the children's event handler change the visibility of
the element that takes the event, in turn the element won't be able
to process any further events.

We don't have the same restrictions with the dragover event.
At that point, it can work independently from the children's.